Change the theme of the website
I'm going to suggest changing the theme of your website to the ["slate" theme](https://pages-themes.github.io/slate/) in order to really grok the difference between markdown, the Github editor, and how it will render in the eventual website.

If you are cool with this, down there ⬇️ there's a button that says something like "Merge and close" - mash that thing and it'll take this branch and merge it into your master branch. If you do, it's good hygiene to delete the branch afterwards (there will be a button for that too). This is how we can collaborate and you can review changes that I make to decide whether you like them or not.